Fixed multi-connection issue
When calling board.deploy(), automate always created a fresh connection,
copied the file and closed the connection again. Calling this in a loop
caused the ssh client to crash at some point. Changed the function
signature to provide an optional connection objec in which case this
connection is used instead of creating a new one.